refactor: move sine-wave synthesis into SoundPlayer, surface engine start outcome
